Fixes some paths, redirects, creates components for turning back into main view
