+++ src/app/page.tsx
@@ -1,20 +1,19 @@
 'use client';
 import 'swiper/css';
-import React, { useRef } from 'react';
-import SwiperContainer from '../components/SwiperContainer';
-import ProductsSegment from '../components/ProductsSegment';
-import SearchField from '../components/SearchField';
-import Header from '@/components/Header';
+import {
+  Header,
+  ProductsSegment,
+  SwiperContainer,
+  SearchField,
+} from '@/components';
 
 const Dashboard = () => {
-  const myRef = useRef(null);
-
   return (
     <section>
       <Header />
       <SearchField />
       <SwiperContainer />
-      <ProductsSegment myRef={myRef} />
+      <ProductsSegment />
     </section>
   );
 };
