+++ src/app/produtos/[segment]/[product]/page.tsx
