+++ src/components/index.tsx
@@ -1,0 +1,7 @@
+import Header from './Header';
+import Product from './Product';
+import ProductsSegment from './ProductsSegment';
+import SearchField from './SearchField';
+import SwiperContainer from './SwiperContainer';
+
+export { Header, Product, ProductsSegment, SearchField, SwiperContainer };
